requirements.txt
fastapi
uvicorn
motor
pydantic[email]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart